Introduce skip buffers
This ensures that we don't recurse into elements that we don't intend
to. It gains us speed and correctness.

Without this concept, if we say try to handle <title> for example, we
might handle <title> at the level we intended to, but also handle one
nested under a separate <image> tag. Adding this optimization skips the
entire <image> tag -- we get to avoid operations we don't care about and
we don't handle tags improperly.